+++ src/main/kotlin/queues/LinkedQueue.kt
@@ -1,4 +1,4 @@
-package queue
+package queues
 
 import linkedlists.DoublyLinkedList
