+++ src/main/kotlin/queues/Queue.kt
@@ -1,4 +1,4 @@
-package queue
+package queues
 
 interface Queue<E> {
     val count: Int
