+++ src/test/kotlin/queues/ArrayQueueTest.kt
@@ -1,4 +1,4 @@
-package queue
+package queues
 
 import org.junit.jupiter.api.Test
