+++ src/test/kotlin/queues/LinkedQueueTest.kt
@@ -1,4 +1,4 @@
-package queue
+package queues
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
